Implemented comment retrieval system! Yay
