+++ routes/comments.js
@@ -1,6 +1,4 @@
 var express = require("express");
-
-var router = express.Router();
 const mysql = require("mysql2");
 
 //here we use the environment variables declared in the .env files to establish a connection to the SQL database
@@ -11,90 +9,15 @@ const options = {
   database: process.env.DB_DATABASE2,
 };
 
-const connection = mysql.createConnection(options);
+const commentsConnection = mysql.createConnection(options);
 
-connection.connect(function (err) {
+commentsConnection.connect(function (err) {
   if (err) {
     console.error("Error connecting: " + err.stack);
     return;
   }
   console.log("Connected to the commentsLikes database!");
 });
-router.post("/postComment", function (req, res, next) {
-  //Use an SQL query to insert the comment into the DB
-  connection.query(process.env.SQL_FOR_INSERTING_COMMENT, [
-    0,
-    null,
-    req.body.commentContent,
-    req.body.dateTime,
-    0,
-    req.body.userPosted,
-  ]);
-
-  //on posting a comment, user is redirected to main page ??
-  res.render("topic1", { user: req.user });
-});
-router.post("/topic1/postComment", function (req, res, next) {
-  //Use an SQL query to insert the comment into the DB
-  connection.query(process.env.SQL_FOR_INSERTING_COMMENT, [
-    1,
-    null,
-    req.body.commentContent,
-    req.body.dateTime,
-    0,
-    req.body.userPosted,
-  ]);
-
-  //on posting a comment, user is redirected to main page ??
-  res.render("topic1", { user: req.user });
-});
-
-router.post("/topic2/postComment", function (req, res, next) {
-  //Use an SQL query to insert the comment into the DB
-  connection.query(process.env.SQL_FOR_INSERTING_COMMENT, [
-    2,
-    null,
-    req.body.commentContent,
-    req.body.dateTime,
-    0,
-    req.body.userPosted,
-  ]);
-
-  //on posting a comment, user is redirected to main page ??
-  res.render("topic2", { user: req.user });
-});
-
-router.post("/topic3/postComment", function (req, res, next) {
-  //Use an SQL query to insert the comment into the DB
-  connection.query(process.env.SQL_FOR_INSERTING_COMMENT, [
-    3,
-    null,
-    req.body.commentContent,
-    req.body.dateTime,
-    0,
-    req.body.userPosted,
-  ]);
-
-  //on posting a comment, user is redirected to main page ??
-  res.render("topic3", { user: req.user });
-});
-
-async function getComments(postID) {
-  const comments = null;
-  connection.query(
-    process.env.SQL_FOR_RETRIEVING_COMMENTS,
-    [postID.toString()],
-    function (err, results) {
-      if (err) {
-        console.log("error getting comments: " + err.stack);
-      }
-      comments = results;
-      //Line B
-      //successCallback(results);
-    }
-  );
-  //Line A
-}
 
 /*
 function getCommentsWrapper(postID) {
@@ -111,9 +34,7 @@ function getCommentsWrapper(postID) {
   });
 }
 */
-
-module.exports = router;
-module.exports = getCommentsWrapper;
+module.exports = commentsConnection;
 
 /* 
 Post format in req.body:
