+++ routes/index.js
@@ -1,55 +1,169 @@
 var express = require("express");
-var getCommentsWrapper = require("./comments");
+var commentsConnection = require("./comments");
 var router = express.Router();
 
 /* GET home page. */
 router.get("/", function (req, res, next) {
-  res.locals.getCommentsWrapper = getCommentsWrapper;
-
   if (!req.user) {
-    res.locals.user = null;
-    res.render("index");
+    commentsConnection.query(
+      process.env.SQL_FOR_RETRIEVING_COMMENTS,
+      [0],
+      function (err, results) {
+        if (err) {
+          console.log("error getting comments: " + err.stack);
+        }
+        res.render("index", { user: null, comments: results });
+      }
+    );
   } else {
-    //this is necessary to define user in index.ejs!!!!
-    res.locals.user = req.user;
-    res.render("index");
+    commentsConnection.query(
+      process.env.SQL_FOR_RETRIEVING_COMMENTS,
+      [0],
+      function (err, results) {
+        if (err) {
+          console.log("error getting comments: " + err.stack);
+        }
+        res.render("index", { user: req.user, comments: results });
+      }
+    );
   }
 });
 
 router.get("/topic1", function (req, res, next) {
-  res.locals.comments = getComments(1);
   if (!req.user) {
-    res.locals.user = null;
-    res.render("topic1");
+    commentsConnection.query(
+      process.env.SQL_FOR_RETRIEVING_COMMENTS,
+      [1],
+      function (err, results) {
+        if (err) {
+          console.log("error getting comments: " + err.stack);
+        }
+        res.render("topc1", { user: null, comments: results });
+      }
+    );
   } else {
-    //this is necessary to define user in index.ejs!!!!
-    res.locals.user = req.user;
-    res.render("topic1");
+    commentsConnection.query(
+      process.env.SQL_FOR_RETRIEVING_COMMENTS,
+      [1],
+      function (err, results) {
+        if (err) {
+          console.log("error getting comments: " + err.stack);
+        }
+        res.render("topic1", { user: req.user, comments: results });
+      }
+    );
   }
 });
 
 router.get("/topic2", function (req, res, next) {
-  res.locals.comments = getComments(2);
   if (!req.user) {
-    res.locals.user = null;
-    res.render("topic2");
+    commentsConnection.query(
+      process.env.SQL_FOR_RETRIEVING_COMMENTS,
+      [2],
+      function (err, results) {
+        if (err) {
+          console.log("error getting comments: " + err.stack);
+        }
+        res.render("topic2", { user: null, comments: results });
+      }
+    );
   } else {
-    //this is necessary to define user in index.ejs!!!!
-    res.locals.user = req.user;
-    res.render("topic2");
+    commentsConnection.query(
+      process.env.SQL_FOR_RETRIEVING_COMMENTS,
+      [2],
+      function (err, results) {
+        if (err) {
+          console.log("error getting comments: " + err.stack);
+        }
+        res.render("topic2", { user: req.user, comments: results });
+      }
+    );
   }
 });
 
 router.get("/topic3", function (req, res, next) {
-  res.locals.comments = getComments(3);
   if (!req.user) {
-    res.locals.user = null;
-    res.render("topic3");
+    commentsConnection.query(
+      process.env.SQL_FOR_RETRIEVING_COMMENTS,
+      [3],
+      function (err, results) {
+        if (err) {
+          console.log("error getting comments: " + err.stack);
+        }
+        res.render("topic3", { user: null, comments: results });
+      }
+    );
   } else {
-    //this is necessary to define user in index.ejs!!!!
-    res.locals.user = req.user;
-    res.render("topic3");
+    commentsConnection.query(
+      process.env.SQL_FOR_RETRIEVING_COMMENTS,
+      [3],
+      function (err, results) {
+        if (err) {
+          console.log("error getting comments: " + err.stack);
+        }
+        res.render("topic3", { user: req.user, comments: results });
+      }
+    );
   }
+});
+
+router.post("/postComment", function (req, res, next) {
+  //Use an SQL query to insert the comment into the DB
+  commentsConnection.query(process.env.SQL_FOR_INSERTING_COMMENT, [
+    0,
+    null,
+    req.body.commentContent,
+    req.body.dateTime,
+    0,
+    req.body.userPosted,
+  ]);
+
+  //on posting a comment, user is redirected to main page ??
+  res.render("index", { user: req.user });
+});
+router.post("/topic1/postComment", function (req, res, next) {
+  //Use an SQL query to insert the comment into the DB
+  commentsConnection.query(process.env.SQL_FOR_INSERTING_COMMENT, [
+    1,
+    null,
+    req.body.commentContent,
+    req.body.dateTime,
+    0,
+    req.body.userPosted,
+  ]);
+
+  //on posting a comment, user is redirected to main page ??
+  res.render("topic1", { user: req.user });
+});
+
+router.post("/topic2/postComment", function (req, res, next) {
+  //Use an SQL query to insert the comment into the DB
+  commentsConnection.query(process.env.SQL_FOR_INSERTING_COMMENT, [
+    2,
+    null,
+    req.body.commentContent,
+    req.body.dateTime,
+    0,
+    req.body.userPosted,
+  ]);
+
+  //on posting a comment, user is redirected to main page ??
+  res.render("topic2", { user: req.user });
+});
+
+router.post("/topic3/postComment", function (req, res, next) {
+  //Use an SQL query to insert the comment into the DB
+  commentsConnection.query(process.env.SQL_FOR_INSERTING_COMMENT, [
+    3,
+    null,
+    req.body.commentContent,
+    req.body.dateTime,
+    0,
+    req.body.userPosted,
+  ]);
+
+  //on posting a comment, user is redirected to main page ??
+  res.render("topic3", { user: req.user });
 });
 
 module.exports = router;
