+++ views/topic1.ejs
@@ -269,48 +269,53 @@
                 <!-- Actual comments  -->
                 <div class="row d-flex justify-content-center">
                     <div class="col-md-12 col-lg-10 col-xl-8">
-                        <div class="card" style="margin-bottom: 20px">
-                            <div class="card-body">
-                                <div>
-                                    <h6 class="fw-bold text-primary mb-1">ElijahsDad</h6>
-                                    <p class="text-muted small mb-0">
-                                        February 69, 2027 00:78 pm
-                                    </p>
-                                </div>
-                                <p class="mt-3 mb-4 pb-2">my dog stepped on a bee :(</p>
-
-                                <div class="d-flex justify-content-end">
-                                    <% if (!user) { %>
-                                        <div class="d-flex">
-                                            <p>3</p>
-                                            <img
-                                                    src="img/fire.png"
-                                                    onclick="alert('guessWhoYourMumSawLastNight')"
-                                            />
-                                        </div>
-                                    <% } else { %>
-                                        <form
-                                                method="POST"
-                                                onsubmit="return false;"
-                                                id=""
-                                                class="d-flex"
-                                        >
-                                            <p class="mb-0 likesCount">4</p>
-                                            <input
-                                                    type="image"
-                                                    src="img/fire.png"
-                                                    class="voteButton inactive"
-                                                    alt="vote"
-                                            />
-                                        </form>
-                                    <% } %>
-                                </div>
+      
+                      <% for (var i = 0; i < comments.length; i++) { %>
+                      <div class="card" style="margin-bottom: 20px">
+                        <div class="card-body">
+                          <div>
+                            <h6 class="fw-bold text-primary mb-1">
+                              <%- comments[i].username %>
+                            </h6>
+                            <p class="text-muted small mb-0">
+                              <%- comments[i].time_posted %>
+                            </p>
+                          </div>
+                          <p class="mt-3 mb-4 pb-2"><%- comments[i].content %></p>
+      
+                          <div class="d-flex justify-content-end">
+                            <% if (!user) { %>
+                            <div class="d-flex">
+                              <p><%- comments[i].likes %></p>
+                              <img
+                                src="img/fire.png"
+                                onclick="alert('guessWhoYourMumSawLastNight')"
+                              />
                             </div>
+                            <% } else { %>
+                            <form
+                              method="POST"
+                              onsubmit="return false;"
+                              id=""
+                              class="d-flex"
+                            >
+                              <p class="mb-0 likesCount"><%- comments[i].likes %></p>
+                              <input
+                                type="image"
+                                src="img/fire.png"
+                                class="voteButton inactive"
+                                alt="vote"
+                              />
+                            </form>
+                            <% } %>
+                          </div>
                         </div>
+                      </div>
+                      <% } %> <%# End of template content %>
                     </div>
                     <!-- This is the end of the comment thread -->
+                  </div>
                 </div>
-            </div>
             <!-- End of comments-->
         </div>
         <!-- End of contents -->
